# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: clipp'n'k API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/clips/{slug}:
    get:
      operationId: getPublicClip
      tags: [public]
      summary: Get a public clip by share slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public clip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicClip"
        "404":
          description: Not found or not public
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PublicClip:
      type: object
      required:
        [
          slug,
          title,
          videoUrl,
          thumbnailUrl,
          width,
          height,
          durationSeconds,
          ownerUsername,
        ]
      properties:
        slug: { type: string }
        title: { type: string }
        videoUrl: { type: string }
        thumbnailUrl: { type: ["string", "null"] }
        width: { type: ["integer", "null"] }
        height: { type: ["integer", "null"] }
        durationSeconds: { type: ["number", "null"] }
        ownerUsername: { type: string }
...